requirements.txt
selenium
slack-sdk
python-dotenv
typing_extensions
